updated multiplexer

+++ main/multiplexer.cpp
@@ -1,9 +1,11 @@
 #include <Wire.h>
 
-void TCA9548A(uint8_t bus){
+void TCA9548A(int bus){
   //Serial.begin(9600);
   Wire.beginTransmission(0x70);  // TCA9548A address
   Wire.write(1 << bus);          // send byte to select bus
   Wire.endTransmission();
-  //Serial.println("Selected bus:" bus);
+  //Serial.println("Selected bus:");
+  //Serial.println(bus);
 }
+
